Refactor post refacto
