+++ front/lib/api/assistant/global_agents/configurations/dust/consts.ts
@@ -1,0 +1,5 @@
+export const DUST_DEEP_NAME = "dust-deep";
+("Deep research with company data, web search/browse, Content Creation, slideshow presentations, and data warehouses.");
+
+export const DUST_DEEP_DESCRIPTION =
+  "Deep research with company data, web search/browse, Content Creation, slideshow presentations, and data warehouses.";
